Upgrade to 7.0.0

+++ axelor-studio/src/main/java/com/axelor/studio/db/repo/MenuBuilderRepo.java
@@ -1,7 +1,7 @@
 /*
  * Axelor Business Solutions
  *
- * Copyright (C) 2020 Axelor (<http://axelor.com>).
+ * Copyright (C) 2021 Axelor (<http://axelor.com>).
  *
  * This program is free software: you can redistribute it and/or  modify
  * it under the terms of the GNU Affero General Public License, version 3,
@@ -41,9 +41,6 @@ public class MenuBuilderRepo extends MenuBuilderRepository {
 
   @Override
   public MenuBuilder save(MenuBuilder menuBuilder) {
-    if (menuBuilder.getName() == null) {
-      menuBuilder.setName("studio-menu-" + menuBuilder.getId());
-    }
     if (menuBuilder.getActionBuilder() != null) {
       menuBuilder.getActionBuilder().setMenuAction(true);
     }
@@ -71,7 +68,10 @@ public class MenuBuilderRepo extends MenuBuilderRepository {
   public void remove(MenuBuilder menuBuilder) {
     MetaMenu metaMenu = menuBuilder.getMetaMenu();
     menuBuilder.setMetaMenu(null);
-    metaService.removeMetaMenu(metaMenu);
+
+    if (metaMenu != null) {
+      metaService.removeMetaMenu(metaMenu);
+    }
 
     ActionBuilder actionBuilder = menuBuilder.getActionBuilder();
